@@ -5,23 +5,21 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Laravel\Sanctum\HasApiTokens;
+use Tymon\JWTAuth\Contracts\JWTSubject;
 
-class User extends Authenticatable
+class User extends Authenticatable implements JWTSubject
 {
-    use HasApiTokens, HasFactory, Notifiable;
-
-    public $incrementing = false;
-    protected $keyType = 'string';
+    use HasFactory, Notifiable;
 
     protected $fillable = [
-        'id',
         'name',
         'email',
         'password',
         'role',
         'permissions',
-        'version',
+        'is_active',
+        'last_login_at',
+        'device_id',
     ];
 
     protected $hidden = [
@@ -29,10 +27,63 @@ class User extends Authenticatable
         'remember_token',
     ];
 
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-        'password' => 'hashed',
-        'permissions' => 'array',
-        'version' => 'integer',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'email_verified_at' => 'datetime',
+            'password' => 'hashed',
+            'permissions' => 'array',
+            'is_active' => 'boolean',
+            'last_login_at' => 'datetime',
+        ];
+    }
+
+    protected static function boot()
+    {
+        parent::boot();
+        
+        static::creating(function ($user) {
+            if (empty($user->uuid)) {
+                $user->uuid = (string) \Illuminate\Support\Str::uuid();
+            }
+        });
+    }
+
+    public function getJWTIdentifier()
+    {
+        return $this->getKey();
+    }
+
+    public function getJWTCustomClaims()
+    {
+        return [
+            'role' => $this->role,
+            'permissions' => $this->permissions,
+        ];
+    }
+
+    public function hasRole(string $role): bool
+    {
+        return $this->role === $role;
+    }
+
+    public function hasPermission(string $permission): bool
+    {
+        return in_array($permission, $this->permissions ?? []);
+    }
+
+    public function suppliers()
+    {
+        return $this->hasMany(Supplier::class, 'created_by');
+    }
+
+    public function collections()
+    {
+        return $this->hasMany(Collection::class, 'collected_by');
+    }
+
+    public function payments()
+    {
+        return $this->hasMany(Payment::class, 'paid_by');
+    }
 }
