@@ -4,37 +4,67 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Support\Str;
 
 class Collection extends Model
 {
-    use SoftDeletes;
-
-    public $incrementing = false;
-    protected $keyType = 'string';
+    use HasFactory, SoftDeletes;
 
     protected $fillable = [
-        'id',
+        'uuid',
         'supplier_id',
         'product_id',
+        'rate_id',
         'quantity',
-        'rate_version_id',
-        'applied_rate',
-        'collection_date',
+        'unit',
+        'rate_at_collection',
+        'total_value',
+        'collected_at',
         'notes',
-        'user_id',
-        'idempotency_key',
-        'version',
+        'sync_status',
+        'metadata',
+        'collected_by',
+        'created_by',
+        'updated_by',
     ];
 
     protected $casts = [
-        'quantity' => 'decimal:2',
-        'applied_rate' => 'decimal:2',
-        'version' => 'integer',
-        'collection_date' => 'datetime',
-        'created_at' => 'datetime',
-        'updated_at' => 'datetime',
-        'deleted_at' => 'datetime',
+        'quantity' => 'decimal:4',
+        'rate_at_collection' => 'decimal:4',
+        'total_value' => 'decimal:2',
+        'collected_at' => 'datetime',
+        'metadata' => 'array',
     ];
+
+    protected static function boot()
+    {
+        parent::boot();
+        
+        static::creating(function ($collection) {
+            if (empty($collection->uuid)) {
+                $collection->uuid = (string) Str::uuid();
+            }
+
+            // Auto-fetch and set rate if not provided
+            if (empty($collection->rate_at_collection) && $collection->product_id) {
+                $rate = $collection->product->getCurrentRate($collection->supplier_id, $collection->collected_at);
+                if ($rate) {
+                    $collection->rate_id = $rate->id;
+                    $collection->rate_at_collection = $rate->rate_value;
+                }
+            }
+
+            // Calculate total value
+            if (!empty($collection->quantity) && !empty($collection->rate_at_collection)) {
+                $collection->total_value = $collection->quantity * $collection->rate_at_collection;
+            }
+        });
+
+        static::saving(function ($collection) {
+            $collection->version++;
+        });
+    }
 
     public function supplier()
     {
@@ -46,13 +76,23 @@ class Collection extends Model
         return $this->belongsTo(Product::class);
     }
 
-    public function rateVersion()
+    public function rate()
     {
-        return $this->belongsTo(RateVersion::class);
+        return $this->belongsTo(Rate::class);
     }
 
-    public function user()
+    public function collector()
     {
-        return $this->belongsTo(User::class);
+        return $this->belongsTo(User::class, 'collected_by');
+    }
+
+    public function creator()
+    {
+        return $this->belongsTo(User::class, 'created_by');
+    }
+
+    public function updater()
+    {
+        return $this->belongsTo(User::class, 'updated_by');
     }
 }
